refactor: move listeners from app.ts -> ezpr/help
