+++ src/ezpr/listeners/submit_ezpr_form.ts
@@ -1,0 +1,25 @@
+import { App, SlackViewAction } from "@slack/bolt";
+import { StringIndexed } from "@slack/bolt/dist/types/helpers";
+
+import { errorOccurred, logger, PublishUsageMetrics } from "../../@lib";
+import { EZPR, EZPR_MODAL_SUBMISSION, VIEW } from "../../constants";
+import { EZPRCommand } from "../cmd";
+import { ParseEZPRFormSubmission } from "../form_submission";
+
+export function registerViewListener(app: App<StringIndexed>) {
+  app.view(EZPR_MODAL_SUBMISSION, async ({ ack, body, client, payload }) => {
+    try {
+      await ack();
+      const args = await ParseEZPRFormSubmission(client, body, payload);
+      const command = new EZPRCommand(client, args);
+      await command.handle();
+      PublishUsageMetrics(VIEW, EZPR, args.numArgs || 0);
+    } catch (error) {
+      const { user } = body as SlackViewAction;
+      if (user !== undefined) {
+        errorOccurred(client, user.id, "", error);
+      }
+      logger.error(error);
+    }
+  });
+}
