+++ src/help/index.ts
@@ -2,4 +2,5 @@ export * from "./";
 export * from "./args";
 export * from "./blocks";
 export * from "./cmd";
+export * from "./listeners";
 export * from "./slash_command";
